Update architecture-overview.md
Added some grammatical and syntactical changes to the Architecture Overview file.

I also think the term Paratime should be used in the same way that Polkadot uses Parachain. That is, the T in Time should not be capitalized, since the T in runtime is not capitalized either. Thus, we can refer to the Paratime concept as a proper noun and then when discussing various paratimes we can use it in the lower case.

Is there any reason why we call it the Oasis Runtime instead of Oasis Paratime? If we are going to develop a new terminology, we should aim to use it consistently.